abstracting error reporter
